requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cytoolz==0.12.0
dnspython==2.2.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
fastapi==0.85.1
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.3.0
httptools==0.5.0
idna==3.4
ipfshttpclient==0.8.0a2
Jinja2==3.1.2
jsonschema==4.17.3
lru-dict==1.1.8
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.3
netaddr==0.8.0
numpy==1.24.2
pandas==1.5.3
parsimonious==0.8.1
protobuf==3.19.5
pycryptodome==3.16.0
pydantic==1.10.2
pymongo==4.3.2
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
redis==4.4.2
requests==2.28.1
rlp==2.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toolz==0.12.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.19.0
varint==1.0.2
watchfiles==0.18.0
web3==5.31.3
websockets==9.1
yarl==1.8.2
